requirements.txt
jupyter>=1.0.0
notebook>=7.0.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-ai-slim[openai]>=0.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
